Move Sprout and Sapling address logic into separate files
